clean up + speed up
